update wasmtest package to sort tests by language
