+++ rwasm/testdata/as-test/src/index.ts
