+++ rwasm/testdata/as-test/src/lib.ts
