+++ rwasm/testdata/as-test/src/tsconfig.json
